Refactor: Move `PasswordConfirmation` interface to `shared.annotations` and update references
