+++ src/main/java/info/mackiewicz/bankapp/presentation/auth/dto/UserRegistrationRequest.java
@@ -1,10 +1,10 @@
 package info.mackiewicz.bankapp.presentation.auth.dto;
 
 import info.mackiewicz.bankapp.shared.annotations.Password;
+import info.mackiewicz.bankapp.shared.annotations.PasswordConfirmation;
 import info.mackiewicz.bankapp.shared.annotations.PasswordMatches;
 import info.mackiewicz.bankapp.shared.annotations.ValidEmail;
 import info.mackiewicz.bankapp.shared.validation.ValidationConstants;
-import info.mackiewicz.bankapp.system.error.handling.dto.interfaces.PasswordConfirmation;
 import info.mackiewicz.bankapp.user.validation.AgeRange;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
