+++ src/main/java/info/mackiewicz/bankapp/presentation/dashboard/dto/ChangePasswordRequest.java
@@ -1,9 +1,9 @@
 package info.mackiewicz.bankapp.presentation.dashboard.dto;
 
 import info.mackiewicz.bankapp.shared.annotations.Password;
+import info.mackiewicz.bankapp.shared.annotations.PasswordConfirmation;
 import info.mackiewicz.bankapp.shared.annotations.PasswordMatches;
 import info.mackiewicz.bankapp.shared.validation.ValidationConstants;
-import info.mackiewicz.bankapp.system.error.handling.dto.interfaces.PasswordConfirmation;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
