+++ src/main/java/info/mackiewicz/bankapp/shared/annotations/PasswordMatchesValidator.java
@@ -1,6 +1,5 @@
 package info.mackiewicz.bankapp.shared.annotations;
 
-import info.mackiewicz.bankapp.system.error.handling.dto.interfaces.PasswordConfirmation;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
