+++ src/main/java/info/mackiewicz/bankapp/system/registration/dto/RegistrationRequest.java
@@ -3,10 +3,10 @@ package info.mackiewicz.bankapp.system.registration.dto;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import info.mackiewicz.bankapp.shared.annotations.Password;
+import info.mackiewicz.bankapp.shared.annotations.PasswordConfirmation;
 import info.mackiewicz.bankapp.shared.annotations.PasswordMatches;
 import info.mackiewicz.bankapp.shared.util.Util;
 import info.mackiewicz.bankapp.shared.validation.ValidationConstants;
-import info.mackiewicz.bankapp.system.error.handling.dto.interfaces.PasswordConfirmation;
 import info.mackiewicz.bankapp.user.model.vo.EmailAddress;
 import info.mackiewicz.bankapp.user.model.vo.Pesel;
 import info.mackiewicz.bankapp.user.model.vo.PhoneNumber;
