+++ src/test/java/info/mackiewicz/bankapp/system/error/handling/core/ApiExceptionHandlerTest.java
@@ -1,11 +1,10 @@
-package info.mackiewicz.bankapp.shared.core;
+package info.mackiewicz.bankapp.system.error.handling.core;
 
-import info.mackiewicz.bankapp.system.error.handling.core.ApiExceptionHandler;
 import info.mackiewicz.bankapp.system.error.handling.core.error.ErrorCode;
-import info.mackiewicz.bankapp.system.error.handling.logger.ApiErrorLogger;
 import info.mackiewicz.bankapp.system.error.handling.dto.BaseApiError;
 import info.mackiewicz.bankapp.system.error.handling.dto.ValidationApiError;
 import info.mackiewicz.bankapp.system.error.handling.dto.ValidationError;
+import info.mackiewicz.bankapp.system.error.handling.logger.ApiErrorLogger;
 import info.mackiewicz.bankapp.system.error.handling.mapping.ApiExceptionToErrorMapper;
 import info.mackiewicz.bankapp.system.error.handling.service.ValidationErrorProcessor;
 import info.mackiewicz.bankapp.system.error.handling.util.RequestUriHandler;
